# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    SeedChain: direct farmer-to-customer agricultural traceability.
    Three roles only: admin, farmer, customer.
    Authentication uses an httpOnly session cookie. Every non-GET request must
    send the header `X-SeedChain-CSRF: 1`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/OrderStatus" } }
      responses:
        "200":
          description: Orders visible to the caller
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Customer places an order; inventory is reserved in the same transaction
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "200":
          description: Duplicate Idempotency-Key; existing order returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "409": { $ref: "#/components/responses/Error" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Order with its event history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/Error" }
  /orders/{id}/{action}:
    post:
      operationId: transitionOrder
      tags: [orders]
      summary: |
        Order state transitions. Farmer: accept, reject, prepare, ready, dispatch, complete.
        Customer: cancel (PENDING/ACCEPTED), confirm-receipt. Admin: cancel, confirm-receipt (with reason).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - name: action
          in: path
          required: true
          schema: { $ref: "#/components/schemas/OrderAction" }
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, maxLength: 100 }
        - name: X-Offline-Replay
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderTransitionBody" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "409": { $ref: "#/components/responses/Error" }
  /orders/{id}/feedback:
    post:
      operationId: submitOrderFeedback
      tags: [orders]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackBody" }
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }

  # ------------------------------------------------------------------ dashboards
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      parameters:
        - name: state
          in: query
          schema: { type: string, enum: [open, all] }
      responses:
        "200":
          description: Alerts (admin = all, farmer = own)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Alert" }
  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Alert" }
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Alert" }

  # ------------------------------------------------------------------ external data
  /market-prices:
    get:
      operationId: getMarketPrices
      tags: [external]
      parameters:
        - { name: state, in: query, schema: { type: string } }
        - { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 90 } }
      responses:
        "200":
          description: Stored external mandi price observations (never fabricated)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketPrices" }
  /weather:
    get:
      operationId: getWeather
      tags: [external]
      parameters:
        - { name: farmId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Latest stored weather observation for the farm location
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeatherReport" }
...
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        details: {}
...
    TraceEvent:
      type: object
      required: [id, lotId, eventType, eventTime, recordedAt, source, metadata]
      properties:
        id: { type: string, format: uuid }
        lotId: { type: string, format: uuid }
        lotCode: { type: string }
        orderId: { type: [string, "null"] }
        eventType: { type: string }
        eventTime: { type: string, format: date-time }
        recordedAt: { type: string, format: date-time }
        actorUserId: { type: [string, "null"] }
        actorName: { type: [string, "null"] }
        actorRole: { type: [string, "null"] }
        location: { type: [string, "null"] }
        latitude: { type: [number, "null"] }
        longitude: { type: [number, "null"] }
        quantityBefore: { type: [number, "null"] }
        quantityChange: { type: [number, "null"] }
        quantityAfter: { type: [number, "null"] }
        status: { type: [string, "null"] }
        reason: { type: [string, "null"] }
        source: { type: string }
        clientEventId: { type: [string, "null"] }
        previousEventId: { type: [string, "null"] }
        correctsEventId: { type: [string, "null"] }
        isPublic: { type: boolean }
        metadata: { type: object, additionalProperties: true }
...
    OrderStatus:
      type: string
      enum: [PENDING, ACCEPTED, REJECTED, PREPARING, READY, DISPATCHED, DELIVERED, CUSTOMER_CONFIRMED, CANCELLED]
    OrderAction:
      type: string
      enum: [accept, reject, prepare, ready, dispatch, complete, cancel, confirm-receipt]
    FulfillmentMethod:
      type: string
      enum: [CUSTOMER_PICKUP, FARMER_DELIVERY, THIRD_PARTY_DELIVERY]
    OrderItem:
      type: object
      required: [id, lotId, lotCode, productName, variety, quantity, unitPrice, lineTotal, unit]
      properties:
        id: { type: string, format: uuid }
        lotId: { type: string, format: uuid }
        lotCode: { type: string }
        productName: { type: string }
        variety: { type: string }
        unit: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        lineTotal: { type: number }
        publicToken: { type: [string, "null"] }
    OrderParty:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: [string, "null"], description: "Visible to the order's farmer and admin only" }
    OrderFarmer:
      type: object
      required: [id, publicName]
      properties:
        id: { type: string, format: uuid }
        publicName: { type: string }
    OrderFeedback:
      type: object
      required: [rating, createdAt]
      properties:
        rating: { type: integer }
        comment: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
    Order:
      type: object
      required: [id, orderCode, status, fulfillmentMethod, totalAmount, items, customer, farmer, createdAt, updatedAt, allowedActions]
      properties:
        id: { type: string, format: uuid }
        orderCode: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        fulfillmentMethod: { $ref: "#/components/schemas/FulfillmentMethod" }
        totalAmount: { type: number }
        deliveryAddress: { type: [string, "null"] }
        customerNotes: { type: [string, "null"] }
        rejectionReason: { type: [string, "null"] }
        cancelReason: { type: [string, "null"] }
        deliveryLocation: { type: [string, "null"] }
        deliveryNotes: { type: [string, "null"] }
        thirdPartyName: { type: [string, "null"] }
        thirdPartyReference: { type: [string, "null"] }
        acceptedAt: { type: [string, "null"], format: date-time }
        preparedAt: { type: [string, "null"], format: date-time }
        readyAt: { type: [string, "null"], format: date-time }
        dispatchedAt: { type: [string, "null"], format: date-time }
        deliveredAt: { type: [string, "null"], format: date-time }
        confirmedAt: { type: [string, "null"], format: date-time }
        cancelledAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        customer: { $ref: "#/components/schemas/OrderParty" }
        farmer: { $ref: "#/components/schemas/OrderFarmer" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        allowedActions:
          type: array
          description: Actions the current caller may perform now (computed by the backend state machine)
          items: { $ref: "#/components/schemas/OrderAction" }
        events:
          type: array
          items: { $ref: "#/components/schemas/TraceEvent" }
        feedback:
          oneOf:
            - $ref: "#/components/schemas/OrderFeedback"
            - type: "null"
    CreateOrderItem:
      type: object
      required: [lotId, quantity]
      additionalProperties: false
      properties:
        lotId: { type: string, format: uuid }
        quantity: { type: number, exclusiveMinimum: 0, maximum: 100000000 }
    CreateOrderBody:
      type: object
      required: [items, fulfillmentMethod]
      additionalProperties: false
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: "#/components/schemas/CreateOrderItem" }
        fulfillmentMethod: { $ref: "#/components/schemas/FulfillmentMethod" }
        deliveryAddress: { type: string, maxLength: 500 }
        customerNotes: { type: string, maxLength: 1000 }
    OrderTransitionBody:
      type: object
      additionalProperties: false
      properties:
        reason: { type: string, maxLength: 1000 }
        eventTime: { type: string, format: date-time }
        deliveryLocation: { type: string, maxLength: 300 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        deliveryNotes: { type: string, maxLength: 1000 }
        thirdPartyName: { type: string, maxLength: 200 }
        thirdPartyReference: { type: string, maxLength: 200 }
        clientEventId: { type: string, format: uuid }
    FeedbackBody:
      type: object
      required: [rating]
      additionalProperties: false
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 1000 }
...
    Alert:
      type: object
      required: [id, type, severity, message, createdAt, metadata]
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        entityType: { type: [string, "null"] }
        entityId: { type: [string, "null"] }
        message: { type: string }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
        acknowledgedAt: { type: [string, "null"], format: date-time }
        acknowledgedBy: { type: [string, "null"] }
        resolvedAt: { type: [string, "null"], format: date-time }
    IngestionRun:
      type: object
      required: [id, source, sourceEndpoint, status, recordCount, processingVersion, startedAt, requestParams]
      properties:
        id: { type: string, format: uuid }
        source: { type: string }
        sourceEndpoint: { type: string }
        requestParams: { type: object, additionalProperties: true }
        processingVersion: { type: string }
        status: { type: string, enum: [RUNNING, SUCCESS, PARTIAL, FAILED, SKIPPED] }
        recordCount: { type: integer }
        error: { type: [string, "null"] }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
    IntegrationStatus:
      type: object
      required: [source, label, configured, recentRuns]
      properties:
        source: { type: string }
        label: { type: string }
        configured: { type: boolean }
        configurationHint: { type: [string, "null"] }
        lastRun:
          oneOf:
            - $ref: "#/components/schemas/IngestionRun"
            - type: "null"
        lastSuccess:
          oneOf:
            - $ref: "#/components/schemas/IngestionRun"
            - type: "null"
        recentRuns: { type: array, items: { $ref: "#/components/schemas/IngestionRun" } }
    MarketPriceObservation:
      type: object
      required: [id, source, sourceEndpoint, observationDate, retrievedAt, state, market, commodity, priceUnit]
      properties:
        id: { type: string, format: uuid }
        source: { type: string }
        sourceEndpoint: { type: string }
        observationDate: { type: string }
        retrievedAt: { type: string, format: date-time }
        state: { type: string }
        district: { type: [string, "null"] }
        market: { type: string }
        commodity: { type: string }
        variety: { type: [string, "null"] }
        grade: { type: [string, "null"] }
        arrivalQuantityTonnes: { type: [number, "null"] }
        minPrice: { type: [number, "null"] }
        maxPrice: { type: [number, "null"] }
        modalPrice: { type: [number, "null"] }
        priceUnit: { type: string }
    MarketPrices:
      type: object
      required: [label, observations, status]
      properties:
        label: { type: string }
        observations: { type: array, items: { $ref: "#/components/schemas/MarketPriceObservation" } }
        status: { $ref: "#/components/schemas/IntegrationStatus" }
    WeatherObservation:
      type: object
      required: [source, sourceEndpoint, observationTime, retrievedAt, latitude, longitude]
      properties:
        source: { type: string }
        sourceEndpoint: { type: string }
        observationTime: { type: string, format: date-time }
        retrievedAt: { type: string, format: date-time }
        latitude: { type: number }
        longitude: { type: number }
        temperatureC: { type: [number, "null"] }
        precipitationMm: { type: [number, "null"] }
        humidityPct: { type: [number, "null"] }
        weatherCode: { type: [integer, "null"] }
        condition: { type: [string, "null"] }
    WeatherReport:
      type: object
      required: [farmId, hasCoordinates, status]
      properties:
        farmId: { type: string, format: uuid }
        hasCoordinates: { type: boolean }
        observation:
          oneOf:
            - $ref: "#/components/schemas/WeatherObservation"
            - type: "null"
        status: { $ref: "#/components/schemas/IntegrationStatus" }